requirements.txt
# === Web Framework ===
fastapi
uvicorn[standard]

# === Markdown, HTTP & Templates ===
markdown
httpx                     # modern async HTTP client
jinja2
cityhash

# === Environment Variables ===
python-dotenv

# === Email ===
aiosmtplib

# === Google Cloud AI & Translation ===
google-cloud-translate
google-api-core
google-genai              # Gemini SDK (already depends on google-api-core)

# === Transformers & Local AI Models ===
transformers
torch                     # backend for transformers
sentence-transformers      # depends on transformers, but needed separately
huggingface_hub
llama-cpp-python           # for local GGUF models (Phi-3, etc.)

# === LangChain Ecosystem ===
langchain

# === Data Validation ===
pydantic                   # used by FastAPI; pin version only if FastAPI compatibility requires it

# === Databases ===
psycopg[binary]             # PostgreSQL driver
asyncpg                     # async PostgreSQL
sqlalchemy                  # ORM
python-arango               # ArangoDB driver
redis                       # caching and stream data

# === Vector Search ===
pgvector

# === Airflow ETL ===
apache-airflow==2.11.0
apache-airflow-providers-postgres
apache-airflow-providers-arangodb

# === Data & ML ===
numpy
pandas
scikit-learn

# === Visualization & Metrics ===
matplotlib
plotly                       # for JSON-serializable chart data
orjson                       # fast JSON serialization
pandas-stubs                 # type hints for Pandas in dev

# === Security & Auth ===
fastapi-keycloak             # for OpenID Connect (Keycloak integration)
python-jose[cryptography]    # JWT handling for auth tokens

# === Performance Monitoring ===
prometheus-fastapi-instrumentator

# Retry + exponential backoff, Logging JSON
tenacity 
structlog